chore: moved API and selector files, added webpack preprocessor, updated import paths
