+++ cypress/e2e/api/update-booking-spec.cy.js
@@ -1,4 +1,4 @@
-import { baseBooking, updatedBooking } from '../../data/booking-data';
+import { baseBooking, updatedBooking } from '@/data/booking-data';
 
 describe('Update booking API Test', () => {
   let bookingId;
