+++ cypress/selectors/inventory-selectors.js
