+++ cypress/selectors/shopping-cart-selectors.js
