+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/AlexMolCode/cypress-automation-tests#readme",
   "devDependencies": {
+    "@cypress/webpack-preprocessor": "^6.0.2",
     "cypress": "^14.0.3",
     "cypress-mochawesome-reporter": "^3.8.2",
     "mochawesome": "^7.1.3",
